refactor: 결제 성공, 실패 이벤트 객체를 common 모듈에 정의된 이벤트 객체로 변경
