--- src/main/java/com/yeoljeong/tripmate/payment/application/event/PaymentCompletedEvent.java
@@ -1,16 +1,0 @@
-package com.yeoljeong.tripmate.payment.application.event;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-import java.util.UUID;
-
-public record PaymentCompletedEvent(
-        UUID eventId,
-        UUID userId,
-        UUID orderId,
-        UUID paymentId,
-        String productName,
-        BigDecimal paymentPrice,
-        Instant paymentDateTime,
-        String paymentMethod
-) { }
