--- src/main/java/com/yeoljeong/tripmate/payment/application/event/PaymentFailedEvent.java
@@ -1,14 +1,0 @@
-package com.yeoljeong.tripmate.payment.application.event;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public record PaymentFailedEvent(
-        UUID eventId,
-        UUID userId,
-        UUID orderId,
-        UUID paymentId,
-        String productName,
-        BigDecimal paymentPrice,
-        String paymentErrorMessage
-) { }
