+++ src/main/java/com/yeoljeong/tripmate/payment/application/port/PaymentEventPublisher.java
@@ -1,7 +1,7 @@
 package com.yeoljeong.tripmate.payment.application.port;
 
-import com.yeoljeong.tripmate.payment.application.event.PaymentCompletedEvent;
-import com.yeoljeong.tripmate.payment.application.event.PaymentFailedEvent;
+import com.yeoljeong.tripmate.event.PaymentCompletedEvent;
+import com.yeoljeong.tripmate.event.PaymentFailedEvent;
 
 public interface PaymentEventPublisher {
     void publish(PaymentCompletedEvent event);
