+++ src/main/java/com/yeoljeong/tripmate/payment/infrastructure/messaging/handler/PaymentEventHandler.java
@@ -1,7 +1,7 @@
 package com.yeoljeong.tripmate.payment.infrastructure.messaging.handler;
 
-import com.yeoljeong.tripmate.payment.application.event.PaymentCompletedEvent;
-import com.yeoljeong.tripmate.payment.application.event.PaymentFailedEvent;
+import com.yeoljeong.tripmate.event.PaymentCompletedEvent;
+import com.yeoljeong.tripmate.event.PaymentFailedEvent;
 import com.yeoljeong.tripmate.payment.application.port.PaymentEventPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
